Update compilation-instructions.md (#828)
There are inconsistent file names in the instructions for compiling binaries: fift and func are mentioned without a directory, while other executables include a directory. Sometimes the directory is prefixed with ./, and other times it’s not. I suggest consistently using the ./ notation, as it follows standard Unix conventions. This change would improve the consistency of the information, make it more reproducible, and speed up the process of mass adoption of the technology.